Add support for exporting values to CSV files
Initially supported is dumping values to the carriers.csv file. When the
dataset contains custom values, a carriers.csv file is created based on
that of the parent dataset, with any custom values added.

When no custom values are present, no file is created at Atlas will use
the parent file instead.